Split Executor header and source files into corresponding class files
