+++ src/aws-cpp-sdk-core/include/aws/core/utils/threading/Executor.h
@@ -4,16 +4,12 @@
  */
 
 #pragma once
+#if !defined(AWS_EXECUTOR_H)
+#define AWS_EXECUTOR_H
 
 #include <aws/core/Core_EXPORTS.h>
-#include <aws/core/utils/memory/stl/AWSQueue.h>
-#include <aws/core/utils/memory/stl/AWSVector.h>
-#include <aws/core/utils/memory/stl/AWSMap.h>
-#include <aws/core/utils/threading/Semaphore.h>
+
 #include <functional>
-#include <future>
-#include <mutex>
-#include <atomic>
 
 namespace Aws
 {
@@ -21,8 +17,6 @@ namespace Aws
     {
         namespace Threading
         {
-            class ThreadTask;
-
             /**
             * Interface for implementing an Executor, to implement a custom thread execution strategy, inherit from this class
             * and override SubmitToThread().
@@ -59,80 +53,11 @@ namespace Aws
                 */
                 virtual bool SubmitToThread(std::function<void()>&&) = 0;
             };
-
-
-            /**
-            * Default Executor implementation. Simply spawns a thread and detaches it.
-            */
-            class AWS_CORE_API DefaultExecutor : public Executor
-            {
-            public:
-                DefaultExecutor() : m_state(State::Free) {}
-                ~DefaultExecutor();
-
-                void WaitUntilStopped() override;
-            protected:
-                enum class State
-                {
-                    Free, Locked, Shutdown
-                };
-                bool SubmitToThread(std::function<void()>&&) override;
-                void Detach(std::thread::id id);
-                std::atomic<State> m_state;
-                Aws::UnorderedMap<std::thread::id, std::thread> m_threads;
-            };
-
-            enum class OverflowPolicy
-            {
-                QUEUE_TASKS_EVENLY_ACROSS_THREADS,
-                REJECT_IMMEDIATELY
-            };
-
-            /**
-            * Thread Pool Executor implementation.
-            */
-            class AWS_CORE_API PooledThreadExecutor : public Executor
-            {
-            public:
-                PooledThreadExecutor(size_t poolSize, OverflowPolicy overflowPolicy = OverflowPolicy::QUEUE_TASKS_EVENLY_ACROSS_THREADS);
-                ~PooledThreadExecutor();
-
-                /**
-                * Rule of 5 stuff.
-                * Don't copy or move
-                */
-                PooledThreadExecutor(const PooledThreadExecutor&) = delete;
-                PooledThreadExecutor& operator =(const PooledThreadExecutor&) = delete;
-                PooledThreadExecutor(PooledThreadExecutor&&) = delete;
-                PooledThreadExecutor& operator =(PooledThreadExecutor&&) = delete;
-
-                /**
-                * Call to ensure the threadpool can be safely destroyed. It blocks until all threads finished.
-                */
-                void WaitUntilStopped() override;
-
-            protected:
-                bool SubmitToThread(std::function<void()>&&) override;
-
-            private:
-                Aws::Queue<std::function<void()>*> m_tasks;
-                mutable std::mutex m_queueLock;
-                Aws::Utils::Threading::Semaphore m_sync;
-                Aws::Vector<ThreadTask*> m_threadTaskHandles;
-                size_t m_poolSize = 0;
-                OverflowPolicy m_overflowPolicy = OverflowPolicy::QUEUE_TASKS_EVENLY_ACROSS_THREADS;
-                bool m_stopped{false};
-
-                /**
-                 * Once you call this, you are responsible for freeing the memory pointed to by task.
-                 */
-                std::function<void()>* PopTask();
-                bool HasTasks() const;
-
-                friend class ThreadTask;
-            };
-
-
         } // namespace Threading
     } // namespace Utils
 } // namespace Aws
+
+// TODO: remove on a next minor API bump from 1.11.x
+#endif // !defined(AWS_EXECUTOR_H)
+#include <aws/core/utils/threading/DefaultExecutor.h>
+#include <aws/core/utils/threading/PooledThreadExecutor.h>
