+++ tests/aws-cpp-sdk-core-tests/utils/threading/ReaderWriterLockTest.cpp
@@ -4,7 +4,7 @@
  */
 
 #include <aws/testing/AwsCppSdkGTestSuite.h>
-#include <aws/core/utils/threading/Executor.h>
+#include <aws/core/utils/threading/DefaultExecutor.h>
 #include <aws/core/utils/threading/ReaderWriterLock.h>
 #include <aws/core/utils/memory/stl/AWSString.h>
